move dir and code
