+++ packages/hono-takibi/src/generators/zod-openapi-hono/openapi/route/params/generate-request-parameter.ts
@@ -3,7 +3,7 @@ import type { Config } from '../../../../../config'
 import { generateParamsObject } from './generate-params-object'
 import { generateRequestParamsArray } from './generate-request-params-array'
 import { generateRequestParams } from './generate-request-params'
-import { generatePropertySchema } from '../../../../zod/generate-zod-property-schema'
+import { generatePropertySchema } from '../../../../zod/property/generate-zod-property-schema'
 import { generateFormatRequestObject } from '../request/object/generate-format-request-object'
 import { generateRequestBody } from '../request/body/generate-request-body'
 import { generateInsertRequestBody } from '../request/body/generate-insert-request-body'
