+++ packages/hono-takibi/src/generators/zod/property/generate-zod-properties-schema.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, it } from 'vitest'
-import type { Schema } from '../../types'
-import type { Config } from '../../config'
+import type { Schema } from '../../../types'
+import type { Config } from '../../../config'
 import { generateZodPropertiesSchema } from './generate-zod-properties-schema'
-import { DEFAULT_CONFIG } from '../../data/test-data'
+import { DEFAULT_CONFIG } from '../../../data/test-data'
 
 const generateZodPropertiesSchemaTestCases: {
   properties: Record<string, Schema>
