+++ packages/hono-takibi/src/generators/zod/property/generate-zod-property-schema.ts
@@ -22,13 +22,13 @@
  * // Returns: 'z.string()'
  */
 
-import type { Schema } from '../../types'
-import type { Config } from '../../config'
-import { isSchemaReference } from '../../core/validator/is-schema-reference'
-import { isArrayWithSchemaReference } from '../../core/validator/is-array-with-schema-reference'
-import { generateReferenceSchema } from './reference/generate-reference-schema'
-import { generateArrayReferenceSchema } from './reference/generate-array-reference-schema'
-import { generateZod } from './generate-zod'
+import type { Schema } from '../../../types'
+import type { Config } from '../../../config'
+import { isSchemaReference } from '../../../core/validator/is-schema-reference'
+import { isArrayWithSchemaReference } from '../../../core/validator/is-array-with-schema-reference'
+import { generateReferenceSchema } from '../reference/generate-reference-schema'
+import { generateArrayReferenceSchema } from '../reference/generate-array-reference-schema'
+import { generateZod } from '../generate-zod'
 
 /**
  * generatePropertySchema
